+++ components/drawer/public-api.ts
@@ -1,0 +1,2 @@
+export * from './nz-drawer.component';
+export * from './nz-drawer.module';