+++ components/ng-zorro-antd.module.ts
@@ -16,6 +16,7 @@ import { NzCheckboxModule } from './checkbox/nz-checkbox.module';
 import { NzCollapseModule } from './collapse/nz-collapse.module';
 import { NzDatePickerModule } from './date-picker/date-picker.module';
 import { NzDividerModule } from './divider/nz-divider.module';
+import { NzDrawerModule } from './drawer/nz-drawer.module';
 import { NzDropDownModule } from './dropdown/nz-dropdown.module';
 import { NzFormModule } from './form/nz-form.module';
 import { NzGridModule } from './grid/nz-grid.module';
@@ -103,6 +104,7 @@ export * from './cascader';
 export * from './tree';
 export * from './tree-select';
 export * from './time-picker';
+export * from './version';
 
 @NgModule({
   exports: [
@@ -141,6 +143,7 @@ export * from './time-picker';
     NzTableModule,
     NzDatePickerModule,
     NzDividerModule,
+    NzDrawerModule,
     NzFormModule,
     NzListModule,
     NzI18nModule,