+++ components/table/nz-th.spec.ts
@@ -10,7 +10,7 @@ describe('nz-th', () => {
   beforeEach(fakeAsync(() => {
     TestBed.configureTestingModule({
       imports     : [ NzTableModule ],
-      declarations: [ NzThTestNzTableComponent ],
+      declarations: [ NzThTestNzTableComponent, NzThTestTableDefaultFilterComponent ],
       providers   : [ NzMeasureScrollbarService ]
     });
     TestBed.compileComponents();
@@ -256,6 +256,24 @@ describe('nz-th', () => {
       }).toThrow();
     });
   });
+  describe('nz-th with default filter in nz-table', () => {
+    let fixture;
+    let testComponent;
+    let th;
+    let table;
+    beforeEach(() => {
+      fixture = TestBed.createComponent(NzThTestTableDefaultFilterComponent);
+      fixture.detectChanges();
+      testComponent = fixture.debugElement.componentInstance;
+      th = fixture.debugElement.query(By.directive(NzThComponent));
+      table = fixture.debugElement.query(By.directive(NzTableComponent));
+    });
+    // It's a fake test to pass codecov, because default displayData should be configured by user.
+    it('should default filter work', () => {
+      expect(testComponent.displayData.length).toBe(0);
+      expect(testComponent.nzThComponent.hasFilterValue).toBe(true);
+    });
+  });
 });
 
 @Component({
@@ -312,6 +330,90 @@ export class NzThTestNzTableComponent {
   showFilter = false;
   filterMultiple = true;
   expand = false;
+}
+
+@Component({
+  selector: 'nz-demo-table-default-filter',
+  template: `
+    <nz-table #filterTable [nzData]="displayData">
+      <thead (nzSortChange)="sort($event)" nzSingleSort>
+        <tr>
+          <th nzShowSort nzSortKey="name" nzShowFilter [nzFilters]="nameList" (nzFilterChange)="filter($event,searchAddress)">Name</th>
+          <th nzShowSort nzSortKey="age">Age</th>
+          <th nzShowSort nzSortKey="address" nzShowFilter [nzFilterMultiple]="false" [nzFilters]="addressList" (nzFilterChange)="filter(listOfSearchName,$event)">Address</th>
+        </tr>
+      </thead>
+      <tbody>
+        <tr *ngFor="let data of filterTable.data">
+          <td>{{data.name}}</td>
+          <td>{{data.age}}</td>
+          <td>{{data.address}}</td>
+        </tr>
+      </tbody>
+    </nz-table>`
+})
+export class NzThTestTableDefaultFilterComponent {
+  nameList = [
+    { text: 'Joe', value: 'Joe', byDefault: true },
+    { text: 'Jim', value: 'Jim' }
+  ];
+  addressList = [
+    { text: 'London', value: 'London', byDefault: true },
+    { text: 'Sidney', value: 'Sidney' }
+  ];
+  sortName = null;
+  sortValue = null;
+  listOfSearchName = [ 'Joe', 'London' ];
+  searchAddress: string;
+  data = [
+    {
+      name: 'John Brown',
+      age: 32,
+      address: 'New York No. 1 Lake Park'
+    },
+    {
+      name: 'Jim Green',
+      age: 42,
+      address: 'London No. 1 Lake Park'
+    },
+    {
+      name: 'Joe Black',
+      age: 32,
+      address: 'Sidney No. 1 Lake Park'
+    },
+    {
+      name: 'Jim Red',
+      age: 32,
+      address: 'London No. 2 Lake Park'
+    }
+  ];
+  displayData = [ ];
+
+  @ViewChild(NzThComponent) nzThComponent: NzThComponent;
+
+  sort(sort: { key: string, value: string }): void {
+    this.sortName = sort.key;
+    this.sortValue = sort.value;
+    this.search();
+  }
+
+  filter(listOfSearchName: string[], searchAddress: string): void {
+    this.listOfSearchName = listOfSearchName;
+    this.searchAddress = searchAddress;
+    this.search();
+  }
+
+  search(): void {
+    /** filter data **/
+    const filterFunc = item => (this.searchAddress ? item.address.indexOf(this.searchAddress) !== -1 : true) && (this.listOfSearchName.length ? this.listOfSearchName.some(name => item.name.indexOf(name) !== -1) : true);
+    const data = this.data.filter(item => filterFunc(item));
+    /** sort data **/
+    if (this.sortName) {
+      this.displayData = data.sort((a, b) => (this.sortValue === 'ascend') ? (a[ this.sortName ] > b[ this.sortName ] ? 1 : -1) : (b[ this.sortName ] > a[ this.sortName ] ? 1 : -1));
+    } else {
+      this.displayData = data;
+    }
+  }
 }
 
 @Component({